version: bump to 0.5.0
* move asu/__init__.py to asu/asu.py
* add CI to publish releases to PyPi.
* apply black style
