+++ requirements.txt
@@ -1,14 +1,18 @@
-Click
-Flask-Cors
-Flask
-Jinja2
-MarkupSafe
-PyNaCl
-Werkzeug
-cffi
-itsdangerous
-pycparser
-redis
-requests
-rq
-six
+certifi==2020.12.5
+cffi==1.14.5
+chardet==4.0.0
+click==7.1.2
+Flask==1.1.2
+Flask-Cors==3.0.10
+idna==2.10
+itsdangerous==1.1.0
+Jinja2==2.11.3
+MarkupSafe==1.1.1
+pycparser==2.20
+PyNaCl==1.4.0
+redis==3.5.3
+requests==2.25.1
+rq==1.8.0
+six==1.15.0
+urllib3==1.26.4
+Werkzeug==1.0.1
